refactor: 로또 자료구조 변경, domain패키치 추가, 테스트 코드 표현 변경(오류 -> 예외)
