+++ src/main/kotlin/lotto/LottoMain.kt
@@ -1,5 +1,7 @@
 package lotto
 
+import lotto.domain.LottoGame
+import lotto.domain.LottoMachine
 import lotto.view.InputView
 import lotto.view.OutputView
