+++ src/main/kotlin/lotto/data/Lotto.kt
@@ -1,5 +1,3 @@
 package lotto.data
 
-import java.util.TreeSet
-
-data class Lotto(val selectNumbers: TreeSet<LottoNumber>)
+data class Lotto(val selectNumbers: LinkedHashSet<LottoNumber>)
